Cache.Supervisor should be rest_for_one
We're monitoring the UrlCache process exits and the only other child is the
simple_one_for_one UrlCacheSupervisor that does nothing but start the
UrlCache processes. The UrlCacheSupervisor needs to be killed and take
the UrlCache workers with it if the Mapper dies but we don't need
one_for_all.