Better local Dockerfile/docker-compose.yml
Don't put .mix and .hex in the app folder as that leads to issues with
permissions and it's a bit weird to have them there.

Instead just create a normal user home directory as well as the app directory.
The app directory stays /opt/app by default to match how bitwalker's production
build Dockerfiles are setup.